feat: add project status to breadcrumbs
